+++ database/migrations/2024_11_22_063321_add_profile_columns.php
@@ -12,12 +12,24 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->string('first_name')->nullable();
-            $table->string('last_name')->nullable();
-            $table->string('gender')->default('male');
-            $table->string('birthday')->nullable();
-            $table->boolean('status_notifications')->default(false);
-            $table->boolean('accept_policy')->default(false);
+            if (!Schema::hasColumn('users', 'first_name')) {
+                $table->string('first_name')->nullable();
+            }
+            if (!Schema::hasColumn('users', 'last_name')) {
+                $table->string('last_name')->nullable();
+            }
+            if (!Schema::hasColumn('users', 'gender')) {
+                $table->string('gender')->default('male');
+            }
+            if (!Schema::hasColumn('users', 'birthday')) {
+                $table->string('birthday')->nullable();
+            }
+            if (!Schema::hasColumn('users', 'status_notifications')) {
+                $table->boolean('status_notifications')->default(false);
+            }
+            if (!Schema::hasColumn('users', 'accept_policy')) {
+                $table->boolean('accept_policy')->default(false);
+            }
         });
     }
 
@@ -27,12 +39,24 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('first_name');
-            $table->dropColumn('last_name');
-            $table->dropColumn('gender');
-            $table->dropColumn('birthday');
-            $table->dropColumn('status_notifications');
-            $table->dropColumn('accept_policy');
+            if (Schema::hasColumn('users', 'first_name')) {
+                $table->dropColumn('first_name');
+            }
+            if (Schema::hasColumn('users', 'last_name')) {
+                $table->dropColumn('last_name');
+            }
+            if (Schema::hasColumn('users', 'gender')) {
+                $table->dropColumn('gender');
+            }
+            if (Schema::hasColumn('users', 'birthday')) {
+                $table->dropColumn('birthday');
+            }
+            if (Schema::hasColumn('users', 'status_notifications')) {
+                $table->dropColumn('status_notifications');
+            }
+            if (Schema::hasColumn('users', 'accept_policy')) {
+                $table->dropColumn('accept_policy');
+            }
         });
     }
 };
